xcode settings for 3-way universal libGoIO_DLL.dylib
To get a 3-way universal (ppc, i386, and x86_64) build of
libGoIO_DLL.dylib working on my 10.6,.4 system I changed 
the following project settings:

Note: Unfortunately xcode project settings are not git-friendly.
When a few simple changes are made huge changes are made in the 
project.pbxproj file. So I've specified in the .gitattributes
file that project.pbxproj should be treated as a 'binary' file.
This means I need to recreate the changes in a narrative form here.

The 10.3 and 1.4 SDKs aren't on my system so I deleted the 
User-Defined project settings from both the project AND
the settings for libGoIO_DLL.dylib.

I changed the Architectures settings for libGoIO_DLL.dylib 
to: "x86_64 ppc i386". The Architectures settings for the 
project are "Standard (32/64-bit Universal)".

I changed the deployment target for the project from
MacOS X 10.3 to MacOS X 10.4. Without doing this I get a 
confusing "_Unwind_Resume" error when building.

There are further changes I needed to make to GUtils_Mac.cpp
to compile libGoIO_DLL.dylib successfully. These are in 
the next commit.