hacks to get libGoIO_DLL.dylib compiling
It's probably easy to fix these problems but I made the
following hacks to get libGoIO_DLL.dylib in a state I
could test it.

SysBeep couldn't be found so it's commented out.

The test for QuickTime produced this error:
  cannot convert 'long int*' to 'SInt32*' for 
  argument '2' to 'OSErr Gestalt(OSType, SInt32*)'
so it's also commented out.

This 'warning' is displayed when using the 64-bit library:

  ../GoIO_DLL_interface.cpp Line: 816

The code makes this assertion (which is untrue when running 64-bit):

  GSTD_ASSERT(sizeof(gtype_int32) == sizeof(long))

Don't know whether this is a problem.