run: ruby extract-dlls.rb, then: goio_ffi.rb
goio_ffi.rb can now be used on different platforms for
testing the vernier goio libraries.

After building the dynamic libraries run this:

  ruby extract-dlls.rb

To generate arch and os-specific library resources 
in dll/

Share the project folder with computers with other
archs or a different os and run from the shared folder:

  ruby goio_ffi.rb

If you are running in JRuby and havce a 64-bit JVM 
you can run in 32 or 64 bit JVMs as follows:

32-bit JVM:

  jruby -J-d32 goio_ffi.rb

64-bit JVM:

  jruby -J-d64 goio_ffi.rb

This will test both the i386 and x86_64 dylibs on the Mac.