commit of Vernier GoIO SDK v2.39
Version 2.39 of the Vernier GoIO SDK was downloaded from this address:
  http://www2.vernier.com/download/GoIO_SDK.zip

Vernier would like to to be able to keep in touch with people using this
SDK and have a form that can be filed out with contact information on
this page: http://www.vernier.com/downloads/gosdk.html

This version of the SDK has a new license. The earlier licenses were in 
effect Public Domain. The license this distribution uses is the simplified
BSD. See license.txt